Memory microoptimizations

+++ VoxelPizza.Base/Collections/BlockStorage.cs
@@ -51,13 +51,13 @@ namespace VoxelPizza.Collections
                     Vector128<uint> v3_32 = Sse2.UnpackLow(v2_16, Vector128<ushort>.Zero).AsUInt32();
                     Vector128<uint> v4_32 = Sse2.UnpackHigh(v2_16, Vector128<ushort>.Zero).AsUInt32();
 
-                    Unsafe.WriteUnaligned(ref Unsafe.Add(ref dst, 0 * Unsafe.SizeOf<Vector128<uint>>()), v1_32);
-                    Unsafe.WriteUnaligned(ref Unsafe.Add(ref dst, 1 * Unsafe.SizeOf<Vector128<uint>>()), v2_32);
-                    Unsafe.WriteUnaligned(ref Unsafe.Add(ref dst, 2 * Unsafe.SizeOf<Vector128<uint>>()), v3_32);
-                    Unsafe.WriteUnaligned(ref Unsafe.Add(ref dst, 3 * Unsafe.SizeOf<Vector128<uint>>()), v4_32);
+                    Unsafe.WriteUnaligned(ref Unsafe.Add(ref dst, 0 * (nuint)Unsafe.SizeOf<Vector128<uint>>()), v1_32);
+                    Unsafe.WriteUnaligned(ref Unsafe.Add(ref dst, 1 * (nuint)Unsafe.SizeOf<Vector128<uint>>()), v2_32);
+                    Unsafe.WriteUnaligned(ref Unsafe.Add(ref dst, 2 * (nuint)Unsafe.SizeOf<Vector128<uint>>()), v3_32);
+                    Unsafe.WriteUnaligned(ref Unsafe.Add(ref dst, 3 * (nuint)Unsafe.SizeOf<Vector128<uint>>()), v4_32);
 
-                    src = ref Unsafe.Add(ref src, 4 * Unsafe.SizeOf<Vector128<byte>>());
-                    dst = ref Unsafe.Add(ref dst, 4 * Unsafe.SizeOf<Vector128<uint>>());
+                    src = ref Unsafe.Add(ref src, 4 * (nuint)Unsafe.SizeOf<Vector128<byte>>());
+                    dst = ref Unsafe.Add(ref dst, 4 * (nuint)Unsafe.SizeOf<Vector128<uint>>());
                 }
 
                 len %= (nuint)Vector128<byte>.Count;
